+++ src/main/java/it/apulia/Esercitazione4/apuliaAirport/flightManagement/FlightServiceImpl.java
@@ -4,22 +4,22 @@ import it.apulia.Esercitazione4.apuliaAirport.bookingmanagement.BookingRepositor
 import it.apulia.Esercitazione4.apuliaAirport.bookingmanagement.PassengerRepository;
 import it.apulia.Esercitazione4.apuliaAirport.bookingmanagement.model.Passeggero;
 import it.apulia.Esercitazione4.apuliaAirport.bookingmanagement.model.Prenotazione;
+import it.apulia.Esercitazione4.apuliaAirport.errors.MyNotAcceptableException;
 import it.apulia.Esercitazione4.apuliaAirport.errors.MyNotFoundException;
 import it.apulia.Esercitazione4.apuliaAirport.flightManagement.model.Tabellone;
 import it.apulia.Esercitazione4.apuliaAirport.flightManagement.model.Volo;
 import it.apulia.Esercitazione4.apuliaAirport.flightManagement.model.VoloOggi;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import java.net.URI;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 
+@Slf4j
 @Service
 public class FlightServiceImpl implements FlightService{
     private final FlightRepository flightRepository;
@@ -34,19 +34,23 @@ public class FlightServiceImpl implements FlightService{
         this.passengerRepository = passengerRepository;
     }
 
-    //TODO add checks
     @Override
     public Volo addVolo(Volo volo) {
-        URI uri = URI.create(ServletUriComponentsBuilder.fromCurrentContextPath()
-                .path("/"+volo.getFlightId()).toUriString());
-        volo.setSelfLink(uri.toString());
-        return this.flightRepository.save(volo);
+        if(!flightRepository.existsById(volo.getFlightId())) {
+            URI uri = URI.create(ServletUriComponentsBuilder.fromCurrentContextPath()
+                    .path("/" + volo.getFlightId()).toUriString());
+            volo.setSelfLink(uri.toString());
+            return this.flightRepository.save(volo);
+        }else
+            throw new MyNotAcceptableException("Il volo con l'id indicato già esiste");
     }
 
-    //TODO add checks
     @Override
     public Volo readVolo(String flightId) {
-        return this.flightRepository.findById(flightId).get();
+        if(flightRepository.existsById(flightId))
+            return this.flightRepository.findById(flightId).get();
+        else
+            throw new MyNotAcceptableException("Il volo con l'id indicato già esiste");
     }
 
     @Override
@@ -54,16 +58,6 @@ public class FlightServiceImpl implements FlightService{
         return flightRepository.findAll();
     }
 
-    /*
-
-    @Override
-    public List<Volo> listAllFlights(String min, String max) {
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy");
-        LocalDate mindate = LocalDate.parse(min,formatter);
-        LocalDate maxdate = LocalDate.parse(max,formatter);
-        return this.flightRepository.findByDepDateAfterAndDepDateBefore(mindate,maxdate);
-    }
-*/
     @Override
     public void updateVolo(Volo volo) {
         if(!this.flightRepository.existsById(volo.getFlightId()))
@@ -74,7 +68,11 @@ public class FlightServiceImpl implements FlightService{
 
     @Override
     public void deleteVolo(String flightId) {
-        this.flightRepository.deleteById(flightId);
+
+        if(!flightRepository.existsById(flightId))
+            log.warn("è stato effettuato un tentativo di eliminazione di un utente non presente all'interno del db");
+        else
+            this.flightRepository.deleteById(flightId);
     }
 
     @Override
@@ -101,6 +99,8 @@ public class FlightServiceImpl implements FlightService{
 
     @Override
     public List<Passeggero> getPassengersFromFlightId(String flightId) {
+        if(!flightRepository.existsById(flightId))
+            throw new MyNotFoundException("Il volo avente l'id da te inserito non esiste");
         List<Prenotazione> listaPrenotazioni = bookingRepository.findByFlightId(flightId);
         List<Passeggero> temp = new ArrayList<Passeggero>();
         listaPrenotazioni.forEach(prenotazione -> {
